requirements.txt
Django==1.6.2
South==0.8.4
django-cors-headers==0.12
djangorestframework==2.3.12
psycopg2==2.5.2
wsgiref==0.1.2
